A couple changes to get SickRage working again
I know absolutely nothing about the code history or innerworkings of
SickBeard/SickRage.  I just used my developer background and stepped
through the code until I found a problem. .. .it seems the latest
SickRage is not liking POST api calls, so I switched it to GET and it
seems to be working.